requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pytest>=7.4.0
pytest-cov>=4.1.0 